Added getShortestPath(startNode,endNode) algorithm
This algorithm finds connects 2 nodes in a way that respects the
arrows and does not go through walls.

This is meant to be the algorithm that will take the nodes returned from DFS
and make a path for the player to traverse them them correctly.

(fixed syntax from last commit so it can compile)